Change the algorithm to add a new feature.
Use a map of components indexed by uid to store the components and
add them to the calendar only at the end of the script.

There is no more the possibility to have duplicate entries since
they are stored into the map. Nonetheless, the IGNORE_DUPLICATE
still makes sense since it indicates to ignore duplicates as
well. The fact is that if an entry is not ignore, it overrides the
previous instead of being added like done previously.

Moreover, the IGNORE_DUPLICATE may be passed the string Latest
that compares the values of the sequence attribute to know which
component is the latest and then should be kept. For instance, if
you record all the ics got by google calendar mail alert, they are
a lot of duplicate, but since it correctly uses the sequence
attribute, it is easy to find out the latest one and to ignore the
previous ones.